backendservice/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
SQLAlchemy==2.0.0
psycopg2-binary==2.9.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.3.0
redis==5.0.0
pycparser==2.22
pydantic_core==2.6.3
PyJWT==2.10.1
app
rq